requirements.txt
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.13
    # via
    #   -r requirements.in
    #   discord-py
aiosignal==1.3.2
    # via aiohttp
attrs==25.3.0
    # via aiohttp
audioop-lts==0.2.1
    # via discord-py
blinker==1.9.0
    # via flask
certifi==2025.6.15
    # via requests
cffi==1.17.1
    # via pynacl
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via flask
discord-ext-voice-recv==0.5.2a179
    # via -r requirements.in
discord-py==2.5.2
    # via
    #   -r requirements.in
    #   discord-ext-voice-recv
flask==3.1.1
    # via -r requirements.in
frozenlist==1.7.0
    # via
    #   aiohttp
    #   aiosignal
gunicorn==23.0.0
    # via -r requirements.in
idna==3.10
    # via
    #   requests
    #   yarl
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
markupsafe==3.0.2
    # via
    #   flask
    #   jinja2
    #   werkzeug
multidict==6.5.0
    # via
    #   aiohttp
    #   yarl
nest-asyncio==1.6.0
    # via -r requirements.in
oauthlib==3.3.0
    # via requests-oauthlib
opuslib==3.0.1
    # via -r requirements.in
packaging==25.0
    # via gunicorn
propcache==0.3.2
    # via
    #   aiohttp
    #   yarl
pycparser==2.22
    # via cffi
pynacl==1.5.0
    # via
    #   -r requirements.in
    #   discord-py
python-dotenv==1.1.0
    # via -r requirements.in
pytube==15.0.0
    # via -r requirements.in
pytz==2025.2
    # via -r requirements.in
requests==2.32.4
    # via requests-oauthlib
requests-oauthlib==2.0.0
    # via -r requirements.in
urllib3==2.5.0
    # via requests
werkzeug==3.1.3
    # via flask
yarl==1.20.1
    # via aiohttp
yt-dlp==2025.6.9
    # via -r requirements.in
